+++ src/index.ts
@@ -1,0 +1,27 @@
+// input
+import { isValidDate } from "./lib/validate/input";
+import { isValidPassword } from "./lib/validate/input";
+import { isValidUrl } from "./lib/validate/input";
+import { isValidEmail } from "./lib/validate/input";
+
+// matchers
+import { isDeepMatch } from "./lib/validate/matchers/isDeepMatch";
+import { isSoftMatch } from "./lib/validate/matchers/isSoftMatch";
+import { isUnique } from "./lib/validate/matchers/isUnique";
+
+// v-listener
+import { VListener } from "./lib/validate/v-listener";
+
+export { VListener }; // v-listener
+export { isValidDate, isValidPassword, isValidUrl, isValidEmail }; // input
+export { isDeepMatch, isSoftMatch, isUnique }; // matchers
+export default {
+  isValidDate,
+  isValidEmail,
+  isValidPassword,
+  isValidUrl,
+  isDeepMatch,
+  isSoftMatch,
+  isUnique,
+  VListener,
+}; // all