+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.3
+
+- Fix where no index file was found.
+- Updated so each folder has its own index file to import from.
+
 # 0.1.0, 0.1.1, 0.1.2
 
 - Initial Setup, copied over from `validatees`.